convert to int first to prevent adding an immense value to offs in case of (unsigned)10 - (unsigned)11 or something similar

+++ src/alsa/device.c
@@ -1112,7 +1112,7 @@ _aaxALSADriverCapture(const void *id, void **data, int offs, size_t *req_frames,
       diff = (float)avail-(float)handle->threshold;
       handle->padding = (handle->padding + diff/(float)no_frames)/2;
       fetch += _MINMAX(roundf(handle->padding), -1, 1);
-      offs += no_frames - fetch;
+      offs += ((int)no_frames - (int)fetch);
 #if 0
 if (roundf(handle->padding))
 printf("avail: %4i (%4i), fetch: %6i\r", avail, handle->threshold, fetch);
